cambios

+++ api/servicio.php
@@ -49,8 +49,40 @@ $con = new Conexion(array(
 
 require "firebase-php-jwt/vendor/autoload.php";
 
+$headers = getallheaders();
+
+$token = "";
+if (isset($headers["Authorization"])) {
+  $token = str_replace("Bearer ", "", $headers["Authorization"]);
+}
+
+try {
+  # el segundo parametro es la clave para codificar y decodificar el JWT
+  # debe ser una string no corta, por eso rellené de guiones
+  $decoded = Firebase\JWT\JWT::decode($token, new Firebase\JWT\Key("Test12345-----------------------------------------------", "HS256"));
+
+  # $usuario puede ser usada para validaciones
+  $usuario = explode("/", $decoded->sub);
+  $id      = $usuario[0];
+  $usr     = $usuario[1];
+  $tipo    = $usuario[2];
+
+  # $login puede ser usada para validaciones
+  $login = true;
+}
+catch (Exception $error) {
+  $usuario = array();
+  $login   = false;
+}
+$esAdmin = $login && $tipo == "1";
+
+# en cada endpoint podemos hacer uso de la variable login para seguridad
+# tenemos la variable login y usuario para realizar más validaciones
+
 ///// PAGOS
-if (isset($_GET ["agre_pagos"])) {
+if (isset($_GET ["agre_pagos"]) && $login) {
+
+
   $prepare = $con->prepare("CALL AgregarPagos(:id_pedido,:monto,:estado_pago,:referencia_paypal, @NUEVOid_pago, @NUEVOid_pedido, @NUEVOmonto, @NUEVOestado_pago, @NUEVOreferencia_paypal)");
   $prepare->bindParam(":id_pedido", $_POST["txtid_pedido"]);
   $prepare->bindParam(":monto", $_POST["txtMonto"]);
@@ -71,7 +103,9 @@ if (isset($_GET ["agre_pagos"])) {
   echo json_encode($pagoAgregado);
 }
 
-elseif (isset($_GET ["pagos"])) {
+elseif (isset($_GET ["pagos"])&& $esAdmin) {
+
+
   $select = $con->select("view_InfoPagos");
   $select->orderBy("id_pago","DESC");
   $select->limit(10);
@@ -81,7 +115,8 @@ elseif (isset($_GET ["pagos"])) {
  
 }
 
-elseif (isset($_GET ["obt_id_pedido"])) {
+elseif (isset($_GET ["obt_id_pedido"])&& $esAdmin) {
+
   $select = $con->select("view_obt_id_pedido");
    $select->orderby("id_pedido","DESC");
   $select->limit(10);
@@ -91,14 +126,14 @@ elseif (isset($_GET ["obt_id_pedido"])) {
 }
 
 ///// PEDIDOS
-elseif (isset($_GET["pedidos"])) {
+elseif (isset($_GET["pedidos"])&& $esAdmin) {
   $select = $con->select("view_pedidos");
   $select->orderby("id_pedido DESC");
 
   header("Content-Type: application/json");
   echo json_encode($select->execute());
 }
-elseif (isset($_GET["pedidosCombo"])) {
+elseif (isset($_GET["pedidosCombo"])&& $esAdmin) {
     $select = $con->select("usuarios", "id_usuario AS value, nombre AS label");
     $select->orderby("nombre ASC");
     $select->limit(10);
@@ -118,7 +153,7 @@ elseif (isset($_GET["pedidosCombo"])) {
     echo json_encode($array);
 }
 
-elseif (isset($_GET["editarPedido"])) {
+elseif (isset($_GET["editarPedido"])&& $esAdmin) {
   $select = $con->select("pedidos", "*");
   $select->where("id_pedido", "=", $_GET["id"]);
 
@@ -127,7 +162,7 @@ elseif (isset($_GET["editarPedido"])) {
 }
 
 
-elseif (isset($_GET["modificarPedido"])) {
+elseif (isset($_GET["modificarPedido"])&& $esAdmin) {
   $prepare = $con->prepare("CALL ModificarPedido(:id_pedido,:id_comprador,:total,:estado)");
   $prepare->bindParam(":id_pedido", $_POST["txtId"]);
   $prepare->bindParam(":id_comprador", $_POST["cboComprador"]);
@@ -142,7 +177,7 @@ elseif (isset($_GET["modificarPedido"])) {
     echo "error";
   }
 }
-elseif(isset($_GET["agreg_pedido"])){
+elseif(isset($_GET["agreg_pedido"]) && $login){
   $prepare = $con->prepare("CALL AgregarPedidoReal(:id_comprador,:total,:estado, @NUEVOid_pedido, @NUEVOid_comprador, @NUEVOtotal, @NUEVOestado)");
   $prepare->bindParam(":id_comprador", $_POST["cboComprador"]);
   $prepare->bindParam(":total", $_POST["txtTotal"]);
@@ -161,7 +196,7 @@ elseif(isset($_GET["agreg_pedido"])){
   echo json_encode($pedidoAgregado);
 }
 ///// PRODUCTOS
-elseif (isset($_GET["productos"])) {
+elseif (isset($_GET["productos"]) && $login) {
   $select = $con->select("view_productos");
   $select->orderby("id DESC");
   $select->limit(10);
@@ -169,7 +204,7 @@ elseif (isset($_GET["productos"])) {
   header("Content-Type: application/json");
   echo json_encode($select->execute());
 }
-elseif (isset($_GET["eliminarProducto"])) {
+elseif (isset($_GET["eliminarProducto"])&& $esAdmin) {
   $prepare = $con->prepare("CALL EliminarProducto(:idProducto)");
     $prepare->bindParam(":idProducto", $_POST["txtId"]);
     $prepare->execute();
@@ -178,7 +213,7 @@ elseif (isset($_GET["eliminarProducto"])) {
     echo "correcto";
 
 }
-elseif (isset($_GET["agregarProducto"])) {
+elseif (isset($_GET["agregarProducto"]) && $login) {
   $nombreImagen = "";
 
   if(isset($_FILES["txtimagen"]) && $_FILES["txtimagen"]["error"] == 0){
@@ -218,7 +253,7 @@ elseif (isset($_GET["agregarProducto"])) {
     echo json_encode($productoAgregado);
 
 }
-elseif (isset($_GET["categoriasCombo"])) {
+elseif (isset($_GET["categoriasCombo"]) && $login) {
     $select = $con->select("categorias", "id_categoria AS value, nombre_categoria AS label");
     $select->orderby("nombre_categoria ASC");
     $select->limit(10);
@@ -237,7 +272,7 @@ elseif (isset($_GET["categoriasCombo"])) {
     header("Content-Type: application/json");
     echo json_encode($array);
 }
-elseif (isset($_GET["vendedorCombo"])) {
+elseif (isset($_GET["vendedorCombo"])&& $login) {
     $select = $con->select("usuarios", "id_usuario AS value, nombre AS label");
     $select->orderby("nombre ASC");
     $select->limit(10);
@@ -259,7 +294,7 @@ elseif (isset($_GET["vendedorCombo"])) {
 
 
 /////DETALLE PEDIDO
-elseif(isset($_GET["detalle_pedido"])) {
+elseif(isset($_GET["detalle_pedido"])&& $esAdmin) {
   $select = $con->select("view_detalle_pedido");
   $select->orderby("id_detalle","DESC");
   $select->limit(10);
@@ -267,7 +302,7 @@ elseif(isset($_GET["detalle_pedido"])) {
   header("Content-Type: application/json");
   echo json_encode($select->execute());
 }
-elseif(isset($_GET["agregardetalle_pedido"])) {
+elseif(isset($_GET["agregardetalle_pedido"]) && $login) {
   $prepare = $con->prepare("CALL AgregarDetalle(:id_pedido,:id_producto,:cantidad,:precio_unitario, @NUEVOid_detalle, @NUEVOid_pedido, @NUEVOid_producto, @NUEVOcantidad, @NUEVOprecio_unitario)");
   $prepare->bindParam(":id_pedido", $_POST["cboPedido"]);
   $prepare->bindParam(":id_producto", $_POST["cboProducto"]);
@@ -284,7 +319,7 @@ elseif(isset($_GET["agregardetalle_pedido"])) {
     echo json_encode($detalleAgregado);
 
 }
-elseif (isset($_GET["PeCombo"])) {
+elseif (isset($_GET["PeCombo"])&& $login) {
     $select = $con->select("pedidos", "id_pedido AS value, id_pedido AS label");
     $select->orderby("id_pedido ASC");
     $select->limit(10);
@@ -303,7 +338,7 @@ elseif (isset($_GET["PeCombo"])) {
     header("Content-Type: application/json");
     echo json_encode($array);
 }
-elseif (isset($_GET["ProCombo"])) {
+elseif (isset($_GET["ProCombo"])&& $login) {
     $select = $con->select("productos", "id_producto AS value, titulo AS label");
     $select->orderby("titulo ASC");
     $select->limit(10);
@@ -322,7 +357,7 @@ elseif (isset($_GET["ProCombo"])) {
     header("Content-Type: application/json");
     echo json_encode($array);
 }
-elseif(isset($_GET["editardetalle_pedido"])) {
+elseif(isset($_GET["editardetalle_pedido"])&& $esAdmin) {
   $select = $con->select("detalle_pedido", "*");
   $select->where("id_detalle", "=", $_GET["id"]);
 
@@ -330,7 +365,7 @@ elseif(isset($_GET["editardetalle_pedido"])) {
   echo json_encode($select->execute());
 
 }
-elseif(isset($_GET["modificardetalle_pedido"])) {
+elseif(isset($_GET["modificardetalle_pedido"])&& $esAdmin) {
   $prepare = $con->prepare("CALL ModificarDetallePedido(:id_detalle,:id_pedido,:id_producto,:cantidad,:precio_unitario)");
   $prepare->bindParam(":id_detalle", $_POST["txtid_detalle"]);
   $prepare->bindParam(":id_pedido", $_POST["cboPedido"]);
@@ -348,7 +383,7 @@ elseif(isset($_GET["modificardetalle_pedido"])) {
 
 }
 
-elseif (isset($_GET["eliminardetalle_pedido"])) {
+elseif (isset($_GET["eliminardetalle_pedido"])&& $esAdmin) {
   $prepare = $con->prepare("CALL eliminarDetalle(:id_detalle)");
     $prepare->bindParam(":id_detalle", $_POST["txtid_detalle"]);
     $prepare->execute();
@@ -360,7 +395,7 @@ elseif (isset($_GET["eliminardetalle_pedido"])) {
 
 
 /////USUARIOS
-elseif (isset($_GET["editarusuario"])){
+elseif (isset($_GET["editarusuario"])&& $esAdmin){
   $select = $con->select("usuarios", "*");
   $select->where("id_usuario", "=", $_GET["id"]);
 
@@ -368,7 +403,7 @@ elseif (isset($_GET["editarusuario"])){
   echo json_encode($select->execute());
 
 }
-elseif (isset($_GET["modificarusuario"])){
+elseif (isset($_GET["modificarusuario"])&& $esAdmin){
   $prepare = $con->prepare("CALL ModificarUsuarioReal(:id_usuario,:nombre,:email,:telefono,:contrasena)");
   $prepare->bindParam(":id_usuario", $_POST["txtIdUsuario"]);
   $prepare->bindParam(":nombre", $_POST["txtNombre"]);
@@ -385,14 +420,14 @@ elseif (isset($_GET["modificarusuario"])){
   }
 
 }
-elseif (isset($_GET["usuarios"])) {
+elseif (isset($_GET["usuarios"])&& $esAdmin) {
   $select = $con->select("view_usuarios");
   $select->orderby("id_usuario DESC");
 
   header("Content-Type: application/json");
   echo json_encode($select->execute());
 }
-elseif (isset($_GET["eliminarUsuario"])) {
+elseif (isset($_GET["eliminarUsuario"])&& $esAdmin) {
   $prepare = $con->prepare("CALL eliminar_usuario(:idUsuario)");
     $prepare->bindParam(":idUsuario", $_POST["txtIdUsuario"]);
     $prepare->execute();
@@ -400,7 +435,7 @@ elseif (isset($_GET["eliminarUsuario"])) {
 
     echo "correcto";
 }
-elseif (isset($_GET["agregarUsuario"])) {
+elseif (isset($_GET["agregarUsuario"])&& $esAdmin) {
   $prepare = $con->prepare("CALL AgregarUsuario(:nombre,:email,:telefono,:contrasena, @NUEVOid_usuario, @NUEVOnombre, @NUEVOemail, @NUEVOtelefono, @NUEVOcontrasena)");
   $prepare->bindParam(":nombre", $_POST["txtNombre"]);
   $prepare->bindParam(":email", $_POST["txtEmail"]);
